ECRI Strategy update

+++ src/StocksCalculator/Models/EcriResult.cs
@@ -6,17 +6,17 @@ namespace StocksCalculator.Models
     public class EcriResult
     {
         public DateTime Date { get; set; }
-        public double EcriLevel { get; set; }
-        public double EcriChange12M { get; set; }
-        public double EcriMovingAvg12 { get; set; }
-        public double EcriMovingAvgMom12 { get; set; }
+        public decimal EcriLevel { get; set; }
+        public decimal EcriChange12M { get; set; }
+        public decimal EcriMovingAvg12 { get; set; }
+        public decimal EcriMovingAvgMom12 { get; set; }
         public byte? CyclePhase { get; set; }
         public byte? CyclePhaseTwoMonthsOld { get; set; }
 
-        public double StocksMovingAvg { get; set; }
-        public double StocksReturn { get; set; }
+        public decimal StocksMovingAvg { get; set; }
+        public decimal StocksReturn { get; set; }
         public bool StocksTrendFollowFilter { get; set; }
-        public double BondsReturn { get; set; }
+        public decimal BondsReturn { get; set; }
 
         public List<AvgReturnByCycle> StocksAvgReturnByCycle { get; set; }
         public List<AvgReturnByCycle> BondsAvgReturnByCycle { get; set; }
@@ -27,7 +27,7 @@ namespace StocksCalculator.Models
     public class AvgReturnByCycle
     {
         public byte CyclePhase { get; set; }
-        public double AvgReturn { get; set; }
+        public decimal AvgReturn { get; set; }
         public bool Result { get; set; }
     }
 }
